Separate spiralized wall from the insets
When spiralize mode is on, the bottom layers of the "vase" will be generated using libArachne
toolpaths, while the spiral wall of the vase will be generated using the old-fashioned offsetted
insets. The difference now is that the spiral inset will only be generated when spiralize is on
and is no longer entangled with the deprecated insets.

CURA-7874