The outer wall toolpath is stitched into a polygon
This is an expensive operation. After a short discussion with
@Ghostkeeper we came to the conclusion that this level of accuracy isn't
 needed. This commit is only for prosperity sake.

We are often assuming that the outer wall_toolpath is the 0th index of
the generated wall_toolpaths after generation, an assert was added, to
check if that is indeed the case.

The stitchContours function is made static and exposed to public,
since this is typically a function that will be used to obtain the
polygon contour of an wall_toolpath in other scenario's. Exposing this
doesn't hurt i.m.o. and will help in the future.

Exposing the width of the ExtrusionLine by taking the width of the first
junction (assuming all junctions have the same width in a line) is also
a valuable addition i.m.o.

Contributes to CURA-7678 and CURA-7880