Reenable Z-seam for extra infill and skin walls.
Mostly just a case of using inset-order optimizer (makes sense, since those walls are insets and they do need optimization of a kind) instead of whatever was used before (mostly just directly dumping the walls to the gcode layer). In order to do that properly, the inset-order-optimizer had to learn how to deal with different wall-types (which comes down to switching what configes are used).

CURA-7712