Adjust relation of threshold to center-dev.-strat.
To refelct the previous situation, the overfill bound is slightly smaller than the underfill one. Otherwise, they relate in the same way to the wall transition threshold, which unlike the previous situation, is about the size of the middle wall being placed (and not the size of the walls surrounding them, be it maximum or minimum).

part of CURA-7686