Fix for too small areas w.r.t. iterative concentric.
Slightly ugly fix/workaround for the fact that boost-polygons still barfs on too small areas. So concentric worked well in case of no gaps (as we could fill it with arachne in one go)but less so where gaps are needed and the iterative method was used, making smaller and smaller input polygons, confusing boost and crashing.

CURA-7148