Correct flipped assertion
I had moved the if-statement that was previously around the assertion into the assertion so that it wouldn't get executed in release. I then also flipped the condition of the if-statement so that it would work correctly. However I had also flipped the condition of the assertion itself then which was not correct.