Use minimum combing boundary rather than preferred combing boundary
If there is a combing path possible, it will make a detour to reduce the amount of retractions and prevent crossing the walls. This is the most major determinant of what the actual travel time is going to be. Whether there is a combing path possible depends on the minimum combing boundary, not the preferred combing boundary, since if there is no preferred combing path possible it will fall back to the minimum combing boundary.
This can cause the planner to slightly underestimate the distance to comb, but that'll be very minimal especially since the combing distance itself is much greater in comparison.

Contributes to issue CURA-7891.