Made outer wall lock default behaviour.
I have been thinking really hard how a beading strategy could be applied
to two different wall regions (inner/outer) and still be true to the
strategy and the difference in wall width specified, and this is
basically the only way I see how.

I think of having the "Lock Wall 0" feature always on (no check-box in
the front-end) just normal behavior.

Rational behind this:

> If two different wall thicknesses are specified for outer and inner
walls, the user wants that outer wall to have that dimension. Since the
outer wall always consist of 1 wall there isn't anything to distribute.
It either is the specified dimension or less if the thickness of the
outline is less then (2 x the optimal outer wall width + 1 x minimal
inner wall width).
Now it is entirely possible that the inner wall has more then 1 walls
(even when 1 inner wall is specified, for the cases of converging walls)
In this case the strategies define the distributed width.

This would result in a cleaner code (since  you basically don't have to
re-distribute all walls according to ratio). Also After running multiple
test prints with a constant outer wall I see a huge improvement in
surface quality.

Contributes to CURA-7979