Remove deviations less than 5 micrometres regardless of length
This used to happen in 4.6 and before. When the algorithm was rewritten this was tightened up. However the Voronoi graph doesn't deal well with long segments with very minor deviation since the Voronoi's coordinates get extremely far out then, which causes floating point rounding errors in the Voronoi's coordinates close to the model. This is causing crashes because the Skeletal Trapezoidation is then unable to find the coordinates that belong to the outline.

Contributes to issue CURA-7970.