Remove scaling to fix self intersections
This was causing the wall placement to become in precise. Rounding errors caused by scaling
down and then up again made the object slightly larger than it was before.

I've tested this with the models that should have been fixed with the original change and
they still seem to be fixed, as well as fixing the newly introduced issue

CURA-7853