Do not wipe for continuously linked paths
The problem is currently that the walls are fed as small segments and
the PathOrderOptimizer ensures they are processed in a correct order and
it really doesn't matter that all segments are small and broken up.
Now all is well, you would think, But then the outer wipe distance comes
in to play. Add the end of a small wall segment it thinks: He, I need to
wipe. I will traverse all long the path a do a travel afterwards. This
occurs after each small wall segment. So when a wall is split in two
walls it does the outer wall wipe thingy.

By storing the end point of the last path and checking if this is
aligned with the start point of the new point I will know it this path
is part of a continuously linked path consisting of smaller segments.
If that is the case the outer wall wipe isn't performed.

Contrinutes to CURA-7949_outer_wall_wipe_distance