Don't replace toolpaths wholesale when infill wall is added.
Due to an oversight, each time an infill wall was needed, the input toolpaths would be reassigned. Because this was an input variable, anything already in there was overwritten. In the context of the current bug (only walls around the dense part of gradual support infill) this solves the part of why not _every_density had walls around it, since the same amount of walls was supposed to be generated for each density blob. This however leaves the second part, that it should only print for the least density (ie only around the support as a whole).

CURA-7934