Wall only around the least dense area of gradual support.
Make it so that if there are dense areas in support (with gradual support infill) walls are only printed around the support as a whole. Take care that denser areas take holes into account as well. For this reason a offset is substracted from the infill overlap for any areas denser than the least dense one.

CURA-7934