Apply also a closing operation when preparing the polygons for arachne
The generation of the SkeletalTrapezoidation does not handle well very thin corners. While in most
cases these corners were properly removed by the closing and opening operation, there were still
some corners left in holes.

These thin corners would lead to the generation of a graph that had edges without twins, which
would then lead crashes and all sorts of problems when generating the beads for the polygons.

This commit fixes that by applying an extra opening and closing operation which makes sure that
very thin corners in holes are properly removed.

Contributes to CURA-7651, CURA-7680, CURA-7948, and CURA-7952