Don't remove any inner walls if there are no inner walls
The range upon which the remove_if function is called would be empty if there are 2 beads. If there is 1 bead or 0 beads, the range would become negative: The start iterator would be beyond the end iterator. This is undefined behaviour. Perhaps this is what is crashing the build made with Clang.

Contributes to issue CURA-8059.