If all else is (epsilon) equal, prefer fixed z-seam location.
Due to fixes in the simplification algorithm, polygonized 'circles' are now have equal corners everywhere. This means that the z-seam --for example, for sharpest corner-- ends up all over the place, since a circle has no real 'sharpes corners'. Instead, fix an arbitraty location as preferred, and only select other z seam locations if the difference to the preference value for that location with the next best one is bigger than a certain epsilon.

CURA-8149