Don't just count last extruder of a layer to see which extruders prime
Previously we were only counting the last extruder of a layer, to see which extruders would need to get primed on that layer. With two extruders this kind of works, because if the extruder was the last one of the layer that means that you had switched to that extruder on that layer (since you switch exactly once per layer). However with more than 2 extruders, this would miss some extruders. It also had the problem that if the extruder order was not optimal (e.g. there are empty extruder plans) it would misbehave, causing travel moves to go to 0,0 in order to prime there.
Conceptually, it doesn't matter in which order extruders are printing on a layer to determine what layer the extruders first need to be primed on. They should just get primed on the first layer at which they are used. Previously this calculation was intertwined with the calculation of extruder order, but I've now put it in a separate function which uses the used extruders per layer to find at which layer an extruder is first used. This not just fixes the bug but also separates the code better to indicate how separate of an operation it really is.

Fixes issue CURA-8188.