Simplify: Fix different output for 'begin' and 'end' of polygon.
Enforce the property that closed polygon loops should have the exact same point as begin and end. Since the beginning isn't moved at all, but the end can, this could lead to a situation where the begin and end points of the polygon are different. This then 'creates' and 'extra' point where there was none, which is very probably still incredibly close to the other point, as none of the simplification constraints where checked, since we assume those two points to be exactly the same (which is in fact how we check if it is indeed a polygon in a few places.

CURA-8275 -- fix for #1449 -- if also a problem for the main branch should be backported probably