Fix getting members of parent class
Since this class is a template class, the code is compiled before some real types are known. The parent class is one of those parts that is still unknown since it is a dependent type. So to fix that, explicitly say that we're using these members from the parent class.

Contributes to issue CURA-7928.