Don't re-print parts of sequences that have already been printed
We track which paths have already been printed in a set. If it's already in the set, then don't print them again. Otherwise, print them and add them to the set.
A set like this is a rather expensive method to add an extra boolean flag to the Path struct... But it does separate concerns better.

Contributes to issue CURA-7928.