Unretract prior to the last travel move before an outer wall
When moving to print an outer wall, the printer should not unretract while the nozzle is already on top of the outer wall. Unretracting there can cause dimensional accuracy issues and visual since the nozzle is sitting on top of the outer wall while unretracting and melting the material, which can leave a blip.

Instead, the printer should unretract before the last travel move when going to print an outer wall. After the unretraction is complete, it can apply the last travel move and start printing the outer wall.

CURA-8365