Unretract on layer change if necessary
When the layer is changed and the next layer is about to start with an outer wall, then we need to ensure that an unretract will happen **before** the last travel move of the previous layer.

This way, the unretract will not happen on top of the outer wall of the next layer, which could cause artifacts.

CURA-8365