Prevent segments from being added to the back of other segments
If a long segment is printed first, and then a different segment is added on later to the back of the long segment (combining two 'legs' into one segment), then there will be a border where the lines are not printed in monotonic order.
To prevent this, we really separate all of the segments including the joints of such segments. Where they join will start a new segment as well. Then each of these segments is printed in monotonic order. That way, no line will be printed before an adjacent line.

Contributes to issue CURA-7928.