Fix edge case failing to detect lines that exactly overlap
There was an edge case not being caught when the two endpoints project to exactly the same points on the perpendicular vector. Since we were using < and > instead of <= and >=, we never detected that case of overlap then.
I've streamlined the conditions a bit. It should be completely accurate now.

Contributes to issue CURA-7928.