Also re-connect halfway through polyline string if it splits
If a single polyline overlaps multiple other polylines, even if it is part of a string of polylines, we must be able to stop there and continue with the rest of the string later, in order to prevent non-monotonic ordering if parts interweave with these polyline strings.

Contributes to issue CURA-7928.