Only find coincident points when using the grid locator
The bucket grid apparently rounds the search space up to whole grid cells. It doesn't actually result in only finding points up to a certain distance. So we'll need to manually filter these results afterwards.
The types get really complicated because of that, but with little possibility of shortening them due to needing to pass them to functions.

Contributes to issue CURA-7928.