Fix broken detection of loops in polyline strings
It was marking it as looping way too often because it was only trying the first adjacent line that wasn't itself. If we make it check all adjacent lines that works better. And it's the same performance, since we needed to check that anyway to detect loops in the first place.

Contributes to issue CURA-7928.