Get farthest endpoint from closest endpoint, not from previous point
If the endpoints are not 100% equal (within the 5um allowed deviation), it could be that the farthest endpoint is the same as the closest endpoint. Instead, we check which is the farthest from the closest endpoint. Since the lines should not be 0um long here, that should always give the opposite endpoint except if the polyline is actually a polygon with closing end. Polygons with closing end should have been marked as such in the start of the algorithm so they shouldn't occur here (and if so, the polyline string just terminates early, nothing terrible).

Contributes to issue CURA-7928.