Limit adjacency in both directions, and allow edge cases
We only want to find adjacent lines if they are adjacent. Due to the polyline chains, we might end up marking a polyline as adjacent even if they are way *before* the current polyline. We now no longer mark anything before the limit as adjacent, but also check if it's before the limit but not too far behind the limit either.
Also, we allow lines to be adjacent even if their endpoints match exactly. This is necessary for polyline chains.

Contributes to issue CURA-7928.