Work in progress on back-pressure compensation.
The formula used right now is likely wildly wrong (because we can't use the formula in the paper 1:1 for one), but can be used to build unit-tests on in parallel. Back-pressure compensation is a way to compensate for the pressure the filament recieves when smooshed up against the previous layer.

CURA-7279