Correct(?) formula to compensate back-pressure.
Plus a description of why this would be correct, given the formula in the paper. However, even if so, that still leaves the question as to wether the interpretation now in the code of 'nominal flow for path (config)', 'nominal line width for path (config)' and 'current/actual (unadjusted) line width of path' are the correct values to plug into that formula.

CURA-7279