Always add as starting point if nothing connects to it yet
If a loose line has no lines connecting to it yet, it needs to be added as a starting point. Otherwise it will not get printed at all. This is regardless of whether it then splits into multiple segments, whether it has only 1 connected line as part of a sequence, or whether it's a single loose line.

Contributes to issue CURA-7928.