Centerdeviation generates centerline
The center deviation strategy now generates a center line again.
When this strategy is used it will either deviate the width
of both inner lines, in the case of an even number of beads, or
deviate the inner line, when there is an uneven number.

It now also handles the splitting and adding of the center line
correctly.

Fixes CURA-8069