Improve assertion for insane point output
A vector with size > 400000 is not really an insane output, considering that even in the S5, which is a 340mmx240mm printer, a vector can end on a point e.g., (335000, 235000). This would give a `vSize` of ~409000 and trigger the "insane output". Also, we have many community printers at the range of ~ 350x350. So if we want to check for an insane output, it's best to check for the actual coordinates.

CURA-8471